Revert "Fix ParacusiaSystem."
This reverts commit 7983bbf85327f7e47d0654b5f0198c68ee17f1b7.
